Move bughunt manifest parsing into module
